refactor. move package of domain object
