--- src/main/java/article/Comment.java
@@ -1,17 +1,0 @@
-package article;
-
-import user.User;
-
-public class Comment {
-    private User writer;
-    private String content;
-    private Article article;
-
-    public Comment(User writer, String content, Article article) {
-        this.writer = writer;
-        this.content = content;
-        this.article = article;
-        article.getComments().add(this);
-    }
-
-}
