+++ src/main/java/domain/User.java
@@ -1,4 +1,4 @@
-package user;
+package domain;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
